# backend/docs/openapi-0.4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThermoTrace API",
    "description": "API do ThermoTrace — monitoramento térmico de remessas por etiqueta NFC.\n\nVersão 0.4: acesso, perfis, cargas e recepção idempotente de evidências. Declarações de início/checkpoint/final preservadas com cadeia de integridade. Receber uma evidência não confirma o hardware nem libera uma carga. Fila Android e decodificação no servidor estão em implementação.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Acesso"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Acesso"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Renovacao"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Acesso"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Acesso"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/plataforma/clientes": {
      "post": {
        "tags": [
          "Clientes e equipe"
        ],
        "summary": "Criar Cliente",
        "operationId": "criar_cliente_api_v1_plataforma_clientes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovoCliente"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Clientes e equipe"
        ],
        "summary": "Listar Clientes",
        "operationId": "listar_clientes_api_v1_plataforma_clientes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usuarios": {
      "get": {
        "tags": [
          "Clientes e equipe"
        ],
        "summary": "Listar Usuarios",
        "operationId": "listar_usuarios_api_v1_usuarios_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clientes e equipe"
        ],
        "summary": "Criar Usuario",
        "operationId": "criar_usuario_api_v1_usuarios_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovoUsuario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usuarios/{usuario_id}/desativar": {
      "post": {
        "tags": [
          "Clientes e equipe"
        ],
        "summary": "Desativar Usuario",
        "operationId": "desativar_usuario_api_v1_usuarios__usuario_id__desativar_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Usuario Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auditoria": {
      "get": {
        "tags": [
          "Clientes e equipe"
        ],
        "summary": "Auditoria",
        "operationId": "auditoria_api_v1_auditoria_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "depois_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Depois Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Login": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "format": "password",
            "title": "Senha",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "title": "Login"
      },
...
      "NovoCliente": {
        "properties": {
          "razao_social": {
            "type": "string",
            "maxLength": 240,
            "minLength": 2,
            "title": "Razao Social"
          },
          "cnpj": {
            "type": "string",
            "pattern": "^[A-Z0-9]{12}[0-9]{2}$",
            "title": "Cnpj"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoEmpresa",
            "default": "embarcador"
          },
          "gestor": {
            "$ref": "#/components/schemas/NovoUsuario"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "razao_social",
          "cnpj",
          "gestor"
        ],
        "title": "NovoCliente"
      },
...
      "NovoUsuario": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "maxLength": 256,
            "minLength": 12,
            "format": "password",
            "title": "Senha",
            "writeOnly": true
          },
          "papel": {
            "$ref": "#/components/schemas/PapelUsuario",
            "default": "operador"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nome",
          "email",
          "senha"
        ],
        "title": "NovoUsuario"
      },
      "PapelUsuario": {
        "type": "string",
        "enum": [
          "operador",
          "gestor",
          "admin"
        ],
        "title": "PapelUsuario"
      },
...
      "Renovacao": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "maxLength": 128,
            "minLength": 40,
            "format": "password",
            "title": "Refresh Token",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "Renovacao"
      },
...
      "TipoEmpresa": {
        "type": "string",
        "enum": [
          "embarcador",
          "transportadora",
          "ambos",
          "plataforma"
        ],
        "title": "TipoEmpresa"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}